pyproject.toml
[project]
name = "gamemanager"
version = "0.1.0"
description = ""
authors = [
    { name = "Yevhenii Melnykov", email = "yevhenii.melnykov@codeit.local" }
]
readme = "README.md"
dependencies = [
    "django>=5.1.7",
    "djangorestframework>=3.15.2",
    "celery>=5.4.0",
    "django-celery-beat>=2.7.0",
    "loguru>=0.7.3",
    "redis>=5.2.1",
    "psycopg[binary]>=3.2.6",
    "uvicorn[standard]>=0.29.0",
    "django-jet-reboot>=1.0.12",
    "drf-spectacular>=0.28.0",
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
    "pydantic>=2.11.3",
    "pytest-freezegun>=0.4.2",
    "factory-boy>=3.3.3",
    "pytest-factoryboy>=2.7.0",
]
requires-python = ">=3.11"

[dependency-groups]
dev = [
    "pytest-cov>=6.0.0",
    "ruff>=0.11.2",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "types-python-dateutil",
    "types-redis"
]

[tool.ruff]
line-length = 120
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".env",
    "env",
    "venv",
    ".venv",
    ".vene-win"
]

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["F405", "F403"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.mypy]
python_version = "3.11"
plugins = []

ignore_missing_imports = true
strict_optional = true
exclude = '(\.venv|\.env|env|venv|\.mypy_cache|\.pytest_cache|\.git)'
